wip improve test for generation
